move dashboard component files
