+++ src/app/_services/main/main.service.ts
@@ -7,11 +7,11 @@ import { BehaviorSubject } from 'rxjs';
 export class MainService {
   public isSmallScreen = new BehaviorSubject(false);
   public readonly routes = [
-    { name: 'dashboard', icon: 'home', link: '/dashboard' },
-    { name: 'play', icon: 'play_arrow', link: '/play' },
+    { name: 'dashboard', icon: 'dashboard', link: '/dashboard' },
+    { name: 'decks', icon: 'content_copy', link: '/decks' },
     { name: 'add', icon: 'add_circle_outline', link: '/add' },
     { name: 'stats', icon: 'bar_chart', link: '/stats' },
-    { name: 'more', icon: 'more_horiz', link: '/more' },
+    // { name: 'more', icon: 'more_horiz', link: '/more' },
   ];
   constructor() {}
 }
