+++ src/app/dashboard/dashboard.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { DashboardRoutingModule } from './dashboard-routing.module';
 import { SharedModule } from '../shared/shared.module';
 import { LayoutModule } from '@angular/cdk/layout';
-import { DashboardComponent } from './dashboard.component';
+import { DashboardComponent } from './components/dashboard/dashboard.component';
 import { MatDividerModule } from '@angular/material/divider';
 import { WordOfTheDayComponent } from './components/word-of-the-day/word-of-the-day.component';
 import { MatButtonModule } from '@angular/material/button';
